fix demo deduplication of code
