+++ packages/web/demo/index.html
@@ -8,18 +8,16 @@
     * { margin: 0; padding: 0; box-sizing: border-box; }
     html, body {
       height: 100%;
-      background-color: #1e1e1e;
+      background-color: #1a1a1a;
       display: flex;
       align-items: center;
       justify-content: center;
     }
     #terminal {
-      font-family: 'Fira Code', 'JetBrains Mono', 'SF Mono', Monaco, Menlo, 'Ubuntu Mono', Consolas, 'Liberation Mono', 'Courier New', monospace;
-      font-size: 14px;
-      line-height: 1.4;
+      width: 80vw;
+      height: 80vh;
+      max-width: 1200px;
       outline: none;
-      letter-spacing: 0;
-      font-variant-ligatures: none;
     }
   </style>
 </head>
